refactor: merge Config.ts and ConfigBase.ts
