+++ packages/client/src/Stream.ts
@@ -17,8 +17,7 @@ import { StreamEndpoints } from './StreamEndpoints'
 import { StreamEndpointsCached } from './StreamEndpointsCached'
 import { EthereumAddress, StreamID, StreamMetadata } from 'streamr-client-protocol'
 import { DEFAULT_PARTITION } from './StreamIDBuilder'
-import { StrictStreamrClientConfig } from './ConfigBase'
-import { ConfigInjectionToken } from './Config'
+import { StrictStreamrClientConfig, ConfigInjectionToken } from './Config'
 import { HttpFetcher } from './utils/HttpFetcher'
 import { PermissionAssignment, PublicPermissionQuery, UserPermissionQuery } from './permission'
